requirements.txt
openai==1.43.0
pynput==1.7.7
pyperclip==1.9.0
python-dotenv==1.0.1
